requirements.txt
asgiref==3.8.1
DateTime==5.5
Django==4.2
django-allauth==0.63.3
django-environ==0.11.2
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0
zope.interface==6.4.post2
python-decouple
